collectors/idfm/requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
google-cloud-bigquery==3.14.1
python-dotenv==1.0.0
